+++ adapters/Facebook/FacebookAdapter/GADFBNativeAdBase.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2019 Google LLC.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,11 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#import <FBAudienceNetwork/FBAudienceNetwork.h>
 #import <Foundation/Foundation.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
-/// Creates and manages Facebook Audience Network banner ads.
-@interface GADFBBannerAd : NSObject
+@interface GADFBNativeAdBase : NSObject
+
+/// Facebook AdChoices view.
+@property(nonatomic, readonly, nonnull) FBAdOptionsView *adOptionsView;
 
 /// Initializes a new instance with |connector| and |adapter|.
 - (nonnull instancetype)initWithGADMAdNetworkConnector:(nonnull id<GADMAdNetworkConnector>)connector
@@ -26,10 +29,11 @@
 /// Unavailable.
 - (nonnull instancetype)init NS_UNAVAILABLE;
 
-/// Starts fetching a banner ad for given |adSize|.
-- (void)getBannerWithSize:(GADAdSize)adSize;
-
-/// Stops the receiver from delegating any notifications from Facebook's Audience Network.
-- (void)stopBeingDelegate;
-
+- (void)loadAdOptionsView;
+- (nullable NSDecimalNumber *)starRating;
+- (nullable NSString *)price;
+- (nullable NSString *)store;
+- (nullable UIView *)adChoicesView;
+- (nullable GADNativeAdImage *)icon;
+- (nullable NSArray *)images;
 @end