+++ adapters/Facebook/FacebookAdapter/GADFBNativeBannerAd.m
@@ -1,4 +1,4 @@
-// Copyright 2019 Google Inc.
+// Copyright 2019 Google LLC.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,43 +12,36 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "GADFBUnifiedNativeAd.h"
-
+#import "GADFBNativeBannerAd.h"
 #import <FBAudienceNetwork/FBAudienceNetwork.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
 #include <stdatomic.h>
-#import "GADFBAdapterDelegate.h"
 #import "GADFBExtraAssets.h"
 #import "GADFBNetworkExtras.h"
 #import "GADFBUtils.h"
 #import "GADMAdapterFacebookConstants.h"
 #import "GADMediationAdapterFacebook.h"
 
-@interface GADFBUnifiedNativeAd () <GADMediatedUnifiedNativeAd,
-                                    FBNativeAdDelegate,
-                                    FBMediaViewDelegate>
+@interface GADFBNativeBannerAd () <GADMediatedUnifiedNativeAd, FBNativeBannerAdDelegate>
 
 @end
 
-@implementation GADFBUnifiedNativeAd {
+@implementation GADFBNativeBannerAd {
   /// Connector from the Google Mobile Ads SDK to receive ad configurations.
   __weak id<GADMAdNetworkConnector> _connector;
 
   /// Adapter for receiving ad request notifications.
   __weak id<GADMAdNetworkAdapter> _adapter;
 
-  /// Native ad obtained from Facebook's Audience Network.
-  FBNativeAd *_nativeAd;
+  /// Native banner ad obtained from Facebook's Audience Network.
+  FBNativeBannerAd *_nativeBannerAd;
 
-  ///  A dictionary of asset names and object pairs for assets that are not handled by properties of
-  ///  the GADMediatedUnifiedNativeAd subclass
+  /// A dictionary of asset names and object pairs for assets that are not handled by properties of
+  /// the GADMediatedUnifiedNativeAd subclass
   NSDictionary<NSString *, id> *_extraAssets;
 
   /// Holds the state for impression being logged.
   atomic_flag _impressionLogged;
-
-  /// Facebook media view.
-  FBMediaView *_mediaView;
 }
 
 /// Empty method to bypass Apple's private method checking since
@@ -57,8 +50,8 @@
 + (void)mediatedNativeAdDidRecordImpression:(id<GADMediatedNativeAd>)mediatedNativeAd {
 }
 
-- (instancetype)initWithGADMAdNetworkConnector:(nonnull id<GADMAdNetworkConnector>)connector
-                                       adapter:(nonnull id<GADMAdNetworkAdapter>)adapter {
+- (nonnull instancetype)initWithGADMAdNetworkConnector:(nonnull id<GADMAdNetworkConnector>)connector
+                                               adapter:(nonnull id<GADMAdNetworkAdapter>)adapter {
   self = [super initWithGADMAdNetworkConnector:connector adapter:adapter];
   if (self) {
     _adapter = adapter;
@@ -67,7 +60,7 @@
   return self;
 }
 
-- (void)requestNativeAd {
+- (void)requestNativeBannerAd {
   id<GADMAdNetworkConnector> strongConnector = _connector;
   id<GADMAdNetworkAdapter> strongAdapter = _adapter;
 
@@ -75,8 +68,8 @@
     return;
   }
 
-  // -[FBNativeAd initWithPlacementID:] throws an NSInvalidArgumentException if the placement ID is
-  // nil.
+  // -[FBNativeBannerAd initWithPlacementID:] throws an NSInvalidArgumentException if the placement
+  // ID is nil.
   NSString *placementID = [strongConnector publisherId];
   if (!placementID) {
     NSError *error = GADFBErrorWithDescription(@"Placement ID cannot be nil.");
@@ -86,26 +79,26 @@
 
   GADFBConfigureMediationService();
 
-  _nativeAd = [[FBNativeAd alloc] initWithPlacementID:placementID];
+  _nativeBannerAd = [[FBNativeBannerAd alloc] initWithPlacementID:placementID];
 
-  if (!_nativeAd) {
+  if (!_nativeBannerAd) {
     NSString *description = [[NSString alloc]
-        initWithFormat:@"Failed to initialize %@.", NSStringFromClass([FBNativeAd class])];
+        initWithFormat:@"Failed to initialize %@.", NSStringFromClass([FBNativeBannerAd class])];
     NSError *error = GADFBErrorWithDescription(description);
     [strongConnector adapter:strongAdapter didFailAd:error];
     return;
   }
-  _nativeAd.delegate = self;
-  [_nativeAd loadAd];
+
+  _nativeBannerAd.delegate = self;
+  [_nativeBannerAd loadAd];
 }
 
 - (void)stopBeingDelegate {
-  _nativeAd.delegate = nil;
-  _mediaView.delegate = nil;
+  _nativeBannerAd.delegate = nil;
 }
 
-- (nullable NSDictionary *)extraAssets {
-  NSString *socialContext = [_nativeAd.socialContext copy];
+- (nullable NSDictionary<NSString *, id> *)extraAssets {
+  NSString *socialContext = [_nativeBannerAd.socialContext copy];
   if (socialContext) {
     return @{GADFBSocialContext : socialContext};
   }
@@ -113,32 +106,28 @@
 }
 
 - (nullable NSString *)headline {
-  return _nativeAd.headline;
+  return _nativeBannerAd.headline;
 }
 
 - (nullable NSString *)advertiser {
-  return _nativeAd.advertiserName;
+  return _nativeBannerAd.advertiserName;
 }
 
 - (nullable NSString *)body {
-  return _nativeAd.bodyText;
+  return _nativeBannerAd.bodyText;
 }
 
 - (nullable NSString *)callToAction {
-  return _nativeAd.callToAction;
+  return _nativeBannerAd.callToAction;
 }
 
 /// Media view.
 - (nullable UIView *)mediaView {
-  return _mediaView;
+  return nil;
 }
 
-/// Returns YES if the ad has video content.
-/// Because the Facebook Audience Network SDK doesn't offer a way to determine whether a native ad
-/// contains a video asset or not, the adapter always returns a MediaView and claims to have video
-/// content.
 - (BOOL)hasVideoContent {
-  return YES;
+  return NO;
 }
 
 #pragma mark - GADMediatedUnifiedNativeAd
@@ -149,78 +138,55 @@
     nonclickableAssetViews:
         (NSDictionary<GADUnifiedNativeAssetIdentifier, UIView *> *)nonclickableAssetViews
             viewController:(UIViewController *)viewController {
-  NSArray *assets = clickableAssetViews.allValues;
+  NSArray<UIView *> *assets = clickableAssetViews.allValues;
   UIImageView *iconView = nil;
   if ([clickableAssetViews[GADUnifiedNativeIconAsset] isKindOfClass:[UIImageView class]]) {
     iconView = (UIImageView *)clickableAssetViews[GADUnifiedNativeIconAsset];
   }
 
-  [_nativeAd registerViewForInteraction:view
-                              mediaView:_mediaView
-                          iconImageView:iconView
-                         viewController:viewController
-                         clickableViews:assets];
+  [_nativeBannerAd registerViewForInteraction:view
+                                iconImageView:iconView
+                               viewController:viewController
+                               clickableViews:assets];
 }
 
 - (void)didUntrackView:(UIView *)view {
-  [_nativeAd unregisterView];
+  [_nativeBannerAd unregisterView];
 }
 
-#pragma mark - FBNativeAdDelegate
+#pragma mark - FBNativeBannerAdDelegate
 
-- (void)nativeAdDidLoad:(nonnull FBNativeAd *)nativeAd {
-  if (_nativeAd) {
-    [_nativeAd unregisterView];
-  }
-  _nativeAd = nativeAd;
-  _mediaView = [[FBMediaView alloc] init];
-  _mediaView.delegate = self;
+- (void)nativeBannerAdDidLoad:(nonnull FBNativeBannerAd *)nativeBannerAd {
   [self loadAdOptionsView];
-  self.adOptionsView.nativeAd = _nativeAd;
-  id<GADMAdNetworkAdapter> strongAdapter = self->_adapter;
-  id<GADMAdNetworkConnector> strongConnector = self->_connector;
+  self.adOptionsView.nativeAd = _nativeBannerAd;
+  id<GADMAdNetworkAdapter> strongAdapter = _adapter;
+  id<GADMAdNetworkConnector> strongConnector = _connector;
   [strongConnector adapter:strongAdapter didReceiveMediatedUnifiedNativeAd:self];
 }
 
-- (void)nativeAdWillLogImpression:(nonnull FBNativeAd *)nativeAd {
+- (void)nativeBannerAdWillLogImpression:(nonnull FBNativeBannerAd *)nativeBannerAd {
   if (atomic_flag_test_and_set(&_impressionLogged)) {
-    GADFB_LOG(@"FBNativeAd is trying to log an impression again. Adapter will ignore "
+    GADFB_LOG(@"FBNativeBannerAd is trying to log an impression again. Adapter will ignore "
               @"duplicate impression pings.");
     return;
   }
+
   [GADMediatedUnifiedNativeAdNotificationSource mediatedNativeAdDidRecordImpression:self];
 }
 
-- (void)nativeAd:(nonnull FBNativeAd *)nativeAd didFailWithError:(nonnull NSError *)error {
+- (void)nativeBannerAd:(nonnull FBNativeBannerAd *)nativeBannerAd
+      didFailWithError:(nonnull NSError *)error {
   id<GADMAdNetworkConnector> strongConnector = _connector;
   id<GADMAdNetworkAdapter> strongAdapter = _adapter;
   [strongConnector adapter:strongAdapter didFailAd:error];
 }
 
-- (void)nativeAdDidClick:(nonnull FBNativeAd *)nativeAd {
+- (void)nativeBannerAdDidClick:(nonnull FBNativeBannerAd *)nativeBannerAd {
   [GADMediatedUnifiedNativeAdNotificationSource mediatedNativeAdDidRecordClick:self];
   [GADMediatedUnifiedNativeAdNotificationSource mediatedNativeAdWillLeaveApplication:self];
 }
 
-- (void)nativeAdDidFinishHandlingClick:(nonnull FBNativeAd *)nativeAd {
-  // Do nothing.
-}
-
-#pragma mark - FBMediaViewDelegate
-
-- (void)mediaViewVideoDidComplete:(nonnull FBMediaView *)mediaView {
-  [GADMediatedUnifiedNativeAdNotificationSource mediatedNativeAdDidEndVideoPlayback:self];
-}
-
-- (void)mediaViewVideoDidPlay:(nonnull FBMediaView *)mediaView {
-  [GADMediatedUnifiedNativeAdNotificationSource mediatedNativeAdDidPlayVideo:self];
-}
-
-- (void)mediaViewVideoDidPause:(nonnull FBMediaView *)mediaView {
-  [GADMediatedUnifiedNativeAdNotificationSource mediatedNativeAdDidPauseVideo:self];
-}
-
-- (void)mediaViewDidLoad:(nonnull FBMediaView *)mediaView {
+- (void)nativeBannerAdDidFinishHandlingClick:(nonnull FBNativeBannerAd *)nativeBannerAd {
   // Do nothing.
 }