+++ adapters/Facebook/FacebookAdapter/GADFBUtils.h
@@ -24,4 +24,14 @@
 
 /// Returns an NSError with NSLocalizedDescriptionKey and NSLocalizedFailureReasonErrorKey values
 /// set to |description|.
-NSError *GADFBErrorWithDescription(NSString *description);
+NSError *_Nonnull GADFBErrorWithDescription(NSString *_Nonnull description);
+
+/// Set the mediation service information in the Facebook Audience Network SDK.
+void GADFBConfigureMediationService(void);
+
+/// Adds |object| to |set| if |object| is not nil.
+void GADMAdapterFacebookMutableSetAddObject(NSMutableSet *_Nullable set, NSObject *_Nonnull object);
+
+/// Sets |value| for |key| in |dictionary| if |value| is not nil.
+void GADMAdapterFacebookMutableDictionarySet(NSMutableDictionary *_Nonnull dictionary,
+                                             id<NSCopying> _Nullable key, id _Nullable value);